+++ src/main/java/uk/ac/rhul/CS3821_GO/View.java
@@ -4,7 +4,7 @@ package uk.ac.rhul.CS3821_GO;
 import java.util.Scanner;
 
 public interface View {
-    void printBoard(int[][] asciiBoard);
+    void printBoard(byte[][] asciiBoard);
     String promptInput(String playerName);
 
     String promptInput(String playerName, String[] scores, Scanner inputBuffer);
