+++ src/test/java/uk/ac/rhul/CS3821_GO/GoASCIIViewTest.java
@@ -35,7 +35,7 @@ class GoASCIIViewTest {
 
     @Test
     void testDimensions(){
-        textView.printBoard(new int[GoModel.BOARD_SIZE_X][GoModel.BOARD_SIZE_Y]);
+        textView.printBoard(new byte[GoModel.BOARD_SIZE_X][GoModel.BOARD_SIZE_Y]);
         String viewOutput = testOutput.toString().replace("\n","").replace("\r","");
         int desiredCharLength = GoModel.BOARD_SIZE_X * GoModel.BOARD_SIZE_Y;
         assertEquals(desiredCharLength, viewOutput.length());
